Add support for Mountain Lion's Notification Center
If Notification Center is available, we use that. Maybe we can add a
setting in the future, but I have a feeling most people on Mountain
Lion use Notification Center.

I removed the feature where if you click the notification, it opens the
URL of the build. I'll add it back once I figure out how to do it in
Notification Center.