Don't sandbox the app
In order to implement auto-updating, we can't have the app
sandboxed. The app needs to be able to write over itself, but
sandboxing prohibits this. There is probably a workaround by adding
the file to the files it can edit, but it's easier to just disable
sandboxing. If we ever want to publish the app to the Mac App Store,
we need to enable sandboxing, but then we can just have a separate
target for that, without Sparkle.